pyproject.toml
[project]
name = "pifacedigital-mqtt-client"
version = "0.1.0"
authors = [
  { name="Erik Petersen", email="eht@it-trans.dk" },
]
description = "MQTT client for..."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "configparser",
    "paho.mqtt",
    "pifacecommon",
    "pifacedigitalio"
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"

[tool.flit.sdist]
include = [
    ".gitignore",
    ".flake8",
    "LICENSE.txt",
    "Makefile",
     "requirements-dev.txt",
    "requirements.txt",
    "etc/",
    "systemd/",
    "tests/",
]

[tool.flit.external-data]
directory = "systemd"

[project.scripts]
pifacedigital-mqtt-client = "pifacedigital_mqtt_client.pifacedigital_mqtt_client:main"
